+++ packages/terra-form-select/src/combobox/Frame.jsx
@@ -43,7 +43,11 @@ const propTypes = {
    */
   intl: intlShape.isRequired,
   /**
-   * Whether the select is in an invalid state.
+   * Whether the select displays as Incomplete. Use when no value has been provided. _(usage note: `required` must also be set)_.
+   */
+  isIncomplete: PropTypes.bool,
+  /**
+   * Whether the select displays as Invalid. Use when value does not meet validation pattern.
    */
   isInvalid: PropTypes.bool,
   /**
@@ -106,6 +110,7 @@ const defaultProps = {
   clearOptionDisplay: undefined,
   disabled: false,
   dropdownAttrs: undefined,
+  isIncomplete: false,
   isInvalid: false,
   isTouchAccessible: false,
   noResultContent: undefined,
@@ -522,7 +527,7 @@ class Frame extends React.Component {
   }
 
   renderToggleButton() {
-    const { intl } = this.props;
+    const { intl, isInvalid } = this.props;
 
     const mobileButtonUsageGuidanceTxt = intl.formatMessage({ id: 'Terra.form.select.mobileButtonUsageGuidance' });
 
@@ -540,19 +545,30 @@ class Frame extends React.Component {
        * prevents users from ever navigating through the select options.
        */
       if (this.state.isInputFocused) {
+        const toggleClasses = cx([
+          'toggle',
+          { 'is-invalid': isInvalid },
+        ]);
+
         return (
-          <div data-terra-form-select-toggle className={cx('toggle')} onMouseDown={this.handleToggleMouseDown}>
+          <div data-terra-form-select-toggle className={toggleClasses} onMouseDown={this.handleToggleMouseDown}>
             <span className={cx('arrow-icon')} />
           </div>
         );
       }
+
+      const toggleClasses = cx([
+        'toggle',
+        'toggle-narrow',
+        { 'is-invalid': isInvalid },
+      ]);
 
       /**
        * Toggle button enables shifting focus to dropdown. This allows iOS users that are using
        * VoiceOver the ability to navigate to the select options.
        */
       return (
-        <div className={cx(['toggle', 'toggle-narrow'])}>
+        <div className={toggleClasses}>
           <button
             type="button"
             className={cx('toggle-btn')}
@@ -566,8 +582,13 @@ class Frame extends React.Component {
       );
     }
 
+    const toggleClasses = cx([
+      'toggle',
+      { 'is-invalid': isInvalid },
+    ]);
+
     return (
-      <div data-terra-form-select-toggle className={cx('toggle')} onMouseDown={this.toggleDropdown}>
+      <div data-terra-form-select-toggle className={toggleClasses} onMouseDown={this.toggleDropdown}>
         <span className={cx('arrow-icon')} />
       </div>
     );
@@ -582,6 +603,7 @@ class Frame extends React.Component {
       display,
       dropdownAttrs,
       intl,
+      isIncomplete,
       isTouchAccessible,
       isInvalid,
       maxHeight,
@@ -604,6 +626,7 @@ class Frame extends React.Component {
       { 'is-disabled': disabled },
       { 'is-focused': this.state.isFocused },
       { 'is-invalid': isInvalid },
+      { 'is-incomplete': (isIncomplete && required && !isInvalid) },
       { 'is-open': this.state.isOpen },
       customProps.className,
     ]);