+++ packages/terra-form-select/src/terra-dev-site/doc/form-select/SearchSelectField.12.doc.jsx
@@ -12,6 +12,8 @@ import SelectFieldExample from '../example/search/SearchSelectField';
 import SelectFieldExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/search/SearchSelectField';
 import RequiredSearchSelectFieldExample from '../example/search/RequiredSearchSelectField';
 import RequiredSearchSelectFieldExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/search/RequiredSearchSelectField';
+import IncompleteSelectFieldExample from '../example/search/IncompleteSearchSelectField';
+import IncompleteSelectFieldExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/search/IncompleteSearchSelectField';
 
 const DocPage = () => (
   <DocTemplate
@@ -34,6 +36,11 @@ const DocPage = () => (
       title: 'Required Search Select Field Example',
       example: <RequiredSearchSelectFieldExample />,
       source: RequiredSearchSelectFieldExampleSrc,
+    }, {
+      title: 'Incomplete Search Select Field Example',
+      description: 'Applies theme-specific styling for incomplete. ***Note: Only use incomplete if given specific guidance, reserved for specific applications when no value has been provided. Not for general use.***',
+      example: <IncompleteSelectFieldExample />,
+      source: IncompleteSelectFieldExampleSrc,
     }]}
   />
 );